requirements.txt
websockets
slacker-asyncio
dispatchonvalue
